Move dialog utils to utils library
