+++ src/libs/dialog/CMakeLists.txt
@@ -2,5 +2,4 @@ STORM_SETUP(
     TARGET_NAME dialog
     TYPE storm_module
     DEPENDENCIES core geometry model renderer sound_service
-    TEST_DEPENDENCIES catch2
 )
